Fix migrations for SQLite bringup
This commit fixes the database migrations such that an SQLite bringup
from empty is now possible.

Note that it was necessary to move the creation of
`newsletter_email_queue.id` column to be at table creation time,
instead of in `2013_10_13_205634_add_fields_to_newsletter_emails`, as
SQLite does not permit the addition of a PRIMARY KEY field to an extant
table.

Therefore, any installations which have run
`2013_10_01_031026_create_newsletter_email_queue_table` will need to
create an auto-incrementing  PRIMARY KEY field named `id` at column index 0
in `newsletter_email_queue`.

It is regrettable to be forced to retroactively change migrations, but
such is life.